+++ src/main/java/com/bonheur/domain/tag/repository/TagRepositoryCustom.java
@@ -2,8 +2,11 @@ package com.bonheur.domain.tag.repository;
 
 import com.bonheur.domain.tag.model.Tag;
 
+import java.util.List;
+
 public interface TagRepositoryCustom {
     Long findOwnTagByTagName(Long memberId, String tagName);
     Tag findOwnTagByTagId(Long memberId, Long tagId);
     Long isExistTag(Long memberId, Long tagId);
+    List<Tag> getTagUsedByMember(Long memberId);
 }
