+++ src/main/java/com/bonheur/domain/tag/repository/TagRepositoryCustomImpl.java
@@ -4,6 +4,8 @@ import com.bonheur.domain.tag.model.Tag;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 
+import java.util.List;
+
 import static com.bonheur.domain.tag.model.QTag.tag;
 import static com.bonheur.domain.membertag.model.QMemberTag.memberTag;
 
@@ -43,5 +45,21 @@ public class TagRepositoryCustomImpl implements TagRepositoryCustom {
                         // tagName에 해당되는 태그
                         tag.id.eq(tagId)
                 ).join(tag.memberTags, memberTag).fetchOne();
+    }
+
+    @Override
+    public List<Tag> getTagUsedByMember(Long memberId) {
+        return queryFactory.select(tag)
+                .from(tag)
+                .leftJoin(tag.memberTags,memberTag).fetchJoin()
+                .leftJoin(memberTag.member).fetchJoin()
+                .where(
+                        memberTag.member.id.eq(memberId)
+                )
+                .distinct()
+                .orderBy(memberTag.updatedAt.desc())
+                .offset(0)
+                .limit(5)
+                .fetch();
     }
 }
