+++ src/main/java/com/bonheur/domain/tag/service/TagService.java
@@ -1,5 +1,6 @@
 package com.bonheur.domain.tag.service;
 
+import com.bonheur.domain.tag.model.dto.GetTagUsedByMemberResponse;
 import com.bonheur.domain.tag.model.dto.CreateTagResponse;
 import com.bonheur.domain.tag.model.dto.GetTagIdResponse;
 
@@ -10,4 +11,5 @@ TagService {
     CreateTagResponse createTags(Long memberId, List<String> tags);
 
     GetTagIdResponse getTagIdByTagName(Long memberId, String tagName);
+    List<GetTagUsedByMemberResponse> getTagUsedByMember(Long memberId);
 }
